transferred checkbox to its own component
