+++ client/components/Checkbox/checkbox.scss
@@ -1,0 +1,11 @@
+.checkbox {
+  cursor: pointer;
+  height: 16px;
+  width: 16px;
+  margin-right: 6px;
+  fill: #878a8c;
+
+  &--checked {
+    fill: $blue;
+  }
+}
