+++ client/pages/subreddits/create/create.scss
@@ -144,12 +144,7 @@
     }
 
     &__adult {
-      &__checkbox {
-        height: 16px;
-        width: 16px;
-        margin-right: 6px;
-        fill: #878a8c;
-      }
+      //
     }
 
     &__button {
